Add normalization and subtraction
The normalization step rationalizes the signs on the numerator and
denominator, as well as dividing top and bottom with the LCM (lowest
common multiple) using a pre-cooked list of prime numbers up to 197.
This limits the size of the num and den to be on the order of 40K.

I've added the primes scripts into the tools sub-directory in case we
need to make the list longer.

The subtraction overload is just a copy and paste of the addition
function with one small obvious change.